pyproject.toml
[tool.poetry]
name = "pt-tutor"
version = "0.1.0"
description = ""
authors = ["Scott Bastek"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.41.0"
openai = "^1.50.1"
langchain = ">=0.3.14,<0.4.0"
pandas = "2.2.3"
st-supabase-connection = "^2.1.0"
wordcloud = "^1.9.4"
python-dotenv = "^1.1.0"
langgraph = "^0.3.21"
typing-extensions = "^4.13.0"
langchain-openai = "^0.3.11"
langchain-core = "^0.3.49"
sounddevice = "^0.5.1"
soundfile = "^0.13.1"
gotrue = "^2.12.4"
google-cloud = "^0.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
